Fonction de visualisation ✅
