+++ app/_app.py
@@ -1,7 +1,7 @@
 from discord import app_commands
 import asyncio
 from config import os, bot, tree, logger,discord, TOKEN
-from tools import help, clear_slash,start_Session,Wait, visualisation
+from tools import help, clear_slash,start_Session,Wait
 import error_embed as embed
 import classement as ldb
 from admin_command import ban
@@ -80,8 +80,8 @@ async def submit(interaction: discord.Interaction, premier: str, deuxieme: str, 
 
 @tree.command(name="visualisation", description="Te montre tes pronos")
 async def visu(interaction : discord.Interaction):
-    #await interaction.response.defer(ephemeral=True)
-    await visualisation(interaction)
+    await interaction.response.defer(ephemeral=True)
+    await pr.visualisation(interaction)
 # _______________________________________________________________________________________________________________________________
 
 @tree.command(name="leaderboard", description="Affiche le clasement des membres")
